+++ project/Characters/Playable/Amazonian/Amazonian.tscn
@@ -2,12 +2,13 @@
 
 [ext_resource path="res://Characters/Playable/Amazonian/Amazonian.png" type="Texture" id=1]
 [ext_resource path="res://Characters/Playable/Amazonian/Amazonian.cs" type="Script" id=2]
-[ext_resource path="res://States/StateComponent.tscn" type="PackedScene" id=3]
+[ext_resource path="res://Components/States/StateComponent.tscn" type="PackedScene" id=3]
 [ext_resource path="res://Combat/Stats.tscn" type="PackedScene" id=4]
 [ext_resource path="res://Combat/CharacterHitbox.tscn" type="PackedScene" id=5]
 [ext_resource path="res://Combat/CharacterHurtBox.tscn" type="PackedScene" id=6]
 [ext_resource path="res://Combat/HpIndicator.tscn" type="PackedScene" id=7]
 [ext_resource path="res://Combat/InvincibilityTimer.tscn" type="PackedScene" id=8]
+
 
 [sub_resource type="Animation" id=1]
 resource_name = "Attack"
