+++ project/Components/States/Character/DeathState.cs
@@ -1,0 +1,14 @@
+﻿namespace DungeonCrawler.Components.States.Character
+{
+    public class DeathState : State<Characters.Character>
+    {
+        public DeathState(Characters.Character node) : base(node)
+        {
+        }
+
+        public override void Init()
+        {
+            Node.Die();
+        }
+    }
+}
