+++ project/Components/States/IState.cs
@@ -1,4 +1,4 @@
-﻿namespace DungeonCrawler.States
+﻿namespace DungeonCrawler.Components.States
 {
     public interface IState
     {
