+++ project/Components/States/StateComponent.tscn
@@ -1,6 +1,7 @@
 [gd_scene load_steps=2 format=2]
 
-[ext_resource path="res://States/StateComponent.cs" type="Script" id=1]
+[ext_resource path="res://Components/States/StateComponent.cs" type="Script" id=1]
+
 
 [node name="StateComponent" type="Node"]
 script = ExtResource( 1 )
