+++ project/Components/Traits/SoftCollision.tscn
@@ -1,6 +1,7 @@
 [gd_scene load_steps=2 format=2]
 
-[ext_resource path="res://Components/SoftCollision.cs" type="Script" id=1]
+[ext_resource path="res://Components/Traits/SoftCollision.cs" type="Script" id=1]
+
 
 [node name="SoftCollision" type="Area2D"]
 collision_layer = 32
