--- project/States/CommonStates/DeathState.cs
@@ -1,17 +1,0 @@
-﻿using DungeonCrawler.Characters;
-using Godot;
-
-namespace DungeonCrawler.States.CommonStates
-{
-    public class DeathState : State<Character>
-    {
-        public DeathState(Character node) : base(node)
-        {
-        }
-
-        public override void Init()
-        {
-            Node.Die();
-        }
-    }
-}
